+++ History.md
@@ -1,2 +1,7 @@
+### 0.3.0
+* remove `data-format`, use filter instead
+* add filter support
+* remove `render` and `event handler` support from model
+
 ### 0.2.3
 * add support for bindings reuse